feat: add Payment Request API support in Boutique and Wallet
Uses Test Wallet as a Payment Handler to process payments made via the
Payment Request API in the Boutique frontend. This includes adding
necessary manifest files, service worker, and UI components to
facilitate the payment flow. The payment handler start page is `/pay`.

The Boutique checkout page actually uses PaymentRequest API to pay via
test wallet. The orders integration is left-over, due to its tight
integration with the existing checkout flow, as it's meant as proof of
concept for the internal Interledger hackathon. But the payment is
actually made, and we get an `incomingPaymentId` back from the wallet in
the PaymentResponse on successful payments.

Works in _Chromium_ based browsers that support Payment Request API and
Payment Handlers. So, won't work in Firefox and Safari as of now. In
Boutique, you won't see the "Payment Handler" button when not supported.

Other wallets can support this as well by adding similar manifest and
service worker files to their frontends (and of course a pay page that
accepts a receiving wallet address and an optional message in
PaymentRequest). The websites that want to support paying with these
wallets mostly only need to add their pay page as a `supportedMethods`
in the PaymentRequest constructor.

---

`payment-request.html` is a simple playground to test the Payment
Request API.

`/pay.tsx` page is largely based on `send.tsx` page, but simplified for
the PaymentHandler use case (the amount, message, recipient address
comes from the PaymentRequest).

The implementation mostly follows the happy-path, with minimal error
handling for brevity.